solution with testOutliers within plotResiduals
testOutliers doesn't work if simulationOutput is not a DHARMa object (it needs the information if integerResponse is TRUE or FALSE. However, plotResiduals allows for vectors in the simulationOutput (ensureDHARMa will create a DHARMA object but without "integerResponse" information). So, first, I test if the. simulationOutput is a vector, and a warning if so. Then I proceed with plotting (or not) the outliers and coloring dots according to the outlier test.